* Adding bigquery client connector helper. * Moving bigquery to gcp sub folder * Adding `get_project_root` helper to help find the source-extractor configs.
