+++ resources/source-extractor/panoptes.conf
@@ -7,7 +7,7 @@
 CATALOG_NAME     test.cat       # name of the output catalog
 CATALOG_TYPE     ASCII_HEAD     # NONE,ASCII,ASCII_HEAD, ASCII_SKYCAT,
                                 # ASCII_VOTABLE, FITS_1.0 or FITS_LDAC
-PARAMETERS_NAME  /var/panoptes/panoptes-pipeline/resources/source-extractor/panoptes.param  # name of the file containing catalog contents
+PARAMETERS_NAME  panoptes.param  # name of the file containing catalog contents
 
 #------------------------------- Extraction ----------------------------------
 
@@ -18,7 +18,7 @@ DETECT_THRESH    3.0            # <sigmas> or <threshold>,<ZP> in mag.arcsec-2
 ANALYSIS_THRESH  1.5            # <sigmas> or <threshold>,<ZP> in mag.arcsec-2
 
 FILTER           Y              # apply filter for detection (Y or N)?
-FILTER_NAME      /usr/share/source-extractor/default.conv   # name of the file containing the filter
+FILTER_NAME      default.conv   # name of the file containing the filter
 
 DEBLEND_NTHRESH  32             # Number of deblending sub-thresholds
 DEBLEND_MINCONT  0.005          # Minimum contrast parameter for deblending
@@ -59,7 +59,7 @@ PIXEL_SCALE      10.3           # size of pixel in arcsec (0=use FITS WCS info)
 #------------------------- Star/Galaxy Separation ----------------------------
 
 SEEING_FWHM      1.5            # stellar FWHM in arcsec
-STARNNW_NAME     /usr/share/source-extractor/default.nnw    # Neural-Network_Weight table filename
+STARNNW_NAME     default.nnw    # Neural-Network_Weight table filename
 
 #------------------------------ Background -----------------------------------
