+++ src/panoptes/pipeline/utils/__init__.py
@@ -1,0 +1,10 @@
+import pkg_resources
+from pathlib import Path
+
+
+def get_project_root() -> Path:
+    return Path(__file__).parent.parent.parent.parent.parent
+
+
+def get_data_file(filename):
+    return pkg_resources.resource_string('panoptes.pipeline', filename)
